+++ R/cliapp.R
@@ -9,6 +9,24 @@ cliapp <- function(theme = getOption("cli.theme"),
     new = function(theme, user_theme, output)
       clii_init(app, theme, user_theme, output),
 
+    ## Meta
+    meta = function(...) {
+      old <- app$output
+      on.exit(app$output <- old, add = TRUE)
+      on.exit(app$signal <- NULL, add = TRUE)
+      out <- rawConnection(raw(1000), open = "w")
+      on.exit(close(out), add = TRUE)
+      app$output <- out
+      app$signal <- FALSE
+
+      for (msg in list(...)) {
+        do.call(app[[msg$type]], msg$args)
+      }
+
+      txt <- rawToChar(rawConnectionValue(out))
+      clii__message(txt, appendLF = FALSE, output = old, signal = TRUE)
+    },
+
     ## Themes
     list_themes = function()
       clii_list_themes(app),